+++ src/prob/test.jl
@@ -492,3 +492,46 @@ function _build_opf_strg_mi(pm::AbstractPowerModel)
         constraint_dcline(pm, i)
     end
 end
+
+
+"opf with tap magnitude and angle as optimization variables"
+function _run_opf_oltc_pst(file, model_type::Type, optimizer; kwargs...)
+    return run_model(file, model_type, optimizer, _build_opf_oltc_pst; kwargs...)
+end
+
+""
+function _build_opf_oltc_pst(pm::AbstractPowerModel)
+    variable_voltage(pm)
+    variable_generation(pm)
+
+    variable_transformer(pm)
+    variable_branch_flow(pm)
+    variable_dcline_flow(pm)
+
+    objective_min_fuel_and_flow_cost(pm)
+
+    constraint_model_voltage(pm)
+
+    for i in ids(pm, :ref_buses)
+        constraint_theta_ref(pm, i)
+    end
+
+    for i in ids(pm, :bus)
+        constraint_power_balance(pm, i)
+    end
+
+    for i in ids(pm, :branch)
+        constraint_ohms_y_oltc_pst_from(pm, i)
+        constraint_ohms_y_oltc_pst_to(pm, i)
+
+        constraint_voltage_angle_difference(pm, i)
+
+        constraint_thermal_limit_from(pm, i)
+        constraint_thermal_limit_to(pm, i)
+    end
+
+    for i in ids(pm, :dcline)
+        constraint_dcline(pm, i)
+    end
+end
+